renaming From: Sergey Zuev <notifications@github.com> Наверное, будет лучше назвать файл store.ts, поскольку в нём описан не один хук
