+++ src/pages/favorite-list.tsx
@@ -1,6 +1,6 @@
 import { useEffect } from 'react';
 import PlaceCard from '../components/place-card';
-import { useAppDispatch, useAppSelector } from '../hooks/use-app-dispatch';
+import { useAppDispatch, useAppSelector } from '../hooks/store';
 import { fetchFavoriteOffersAction } from '../store/api-actions';
 import { OfferEntity } from '../types';
